+++ 1.21.9/src/main/java/com/bleudev/ppl_utils/PepelandUtils.java
@@ -1,8 +1,8 @@
 package com.bleudev.ppl_utils;
 
 import com.bleudev.ppl_utils.config.PplUtilsConfig;
-import com.bleudev.ppl_utils.custom.Keys;
-import com.bleudev.ppl_utils.custom.debug.hud.WorldBorderDebugHudEntry;
+import com.bleudev.ppl_utils.custom.PepelandUtilsDebugHudEntries;
+import com.bleudev.ppl_utils.custom.PepelandUtilsKeyBindings;
 import com.bleudev.ppl_utils.util.ServerUtils;
 import com.bleudev.ppl_utils.util.helper.DiamondHelper;
 import com.bleudev.ppl_utils.util.helper.ErrorScreenHelper;
@@ -16,7 +16,6 @@ import net.fabricmc.fabric.api.client.rendering.v1.hud.VanillaHudElements;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gl.RenderPipelines;
 import net.minecraft.client.gui.DrawContext;
-import net.minecraft.client.gui.hud.debug.DebugHudEntries;
 import net.minecraft.client.gui.screen.ChatScreen;
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
 import net.minecraft.client.render.RenderTickCounter;
@@ -46,7 +45,8 @@ public class PepelandUtils implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         PplUtilsConfig.initialize();
-        Keys.initialize();
+        PepelandUtilsKeyBindings.initialize();
+        PepelandUtilsDebugHudEntries.initialize();
 
         // Initialize data storage
         DataStorageHelper.load();
@@ -59,9 +59,6 @@ public class PepelandUtils implements ClientModInitializer {
         restartHelper = new RestartHelper();
         GlobalChatHelper.INSTANCE = new GlobalChatHelper(false);
         ErrorScreenHelper.INSTANCE = new ErrorScreenHelper();
-
-        LOGGER.debug("Register {} debug hud entry", getIdentifier("world_border"));
-        DebugHudEntries.register(getIdentifier("world_border"), new WorldBorderDebugHudEntry());
 
         ClientPlayConnectionEvents.JOIN.register((handler, sender, client) -> {
             LOGGER.info("Try send beta mode message");
@@ -81,20 +78,20 @@ public class PepelandUtils implements ClientModInitializer {
         ClientTickEvents.END_CLIENT_TICK.register(client -> {
             if (beta_mode_message_ticks > 0) beta_mode_message_ticks--;
 
-            while (Keys.LOBBY_KEY.wasPressed()) executeLobby(client);
-            while (Keys.SIT_KEY.wasPressed()) executeSit(client);
-            while (Keys.LAY_KEY.wasPressed()) executeLay(client);
-            while (Keys.SEND_TO_GLOBAL_CHAT_KEY.wasPressed())
+            while (PepelandUtilsKeyBindings.LOBBY_KEY.wasPressed()) executeLobby(client);
+            while (PepelandUtilsKeyBindings.SIT_KEY.wasPressed()) executeSit(client);
+            while (PepelandUtilsKeyBindings.LAY_KEY.wasPressed()) executeLay(client);
+            while (PepelandUtilsKeyBindings.SEND_TO_GLOBAL_CHAT_KEY.wasPressed())
                 if (isGlobalChatWorking(client))
                     client.setScreen(new ChatScreen("/" + GLOBAL_CHAT_COMMAND + " ", false));
-            while (Keys.TOGGLE_GLOBAL_CHAT_KEY.wasPressed()) {
+            while (PepelandUtilsKeyBindings.TOGGLE_GLOBAL_CHAT_KEY.wasPressed()) {
                 if (isGlobalChatWorking(client)) {
                     GlobalChatHelper.INSTANCE.toggle();
                     GlobalChatHelper.INSTANCE.sendToggleMessage(client);
                 } else GlobalChatHelper.INSTANCE.sendToggleErrorMessage(client);
             }
             if (client.player == null) return;
-            while (Keys.SHOW_PING.wasPressed()) {
+            while (PepelandUtilsKeyBindings.SHOW_PING.wasPressed()) {
                 var p = ServerUtils.getPing(client);
                 if (p == -1) client.player.sendMessage(Text.translatable("ppl_utils.text.show_ping.failure").formatted(Formatting.RED), true);
                 else client.player.sendMessage(Text.translatable("ppl_utils.text.show_ping.success").append(Text.translatable("ppl_utils.text.general.ping", p).formatted(Formatting.GREEN)), true);