+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 ## New features
 [+] New lobby button style: arrow
+
+## Fixes
+[/] Fix world border debug hud entry
+
+## Dev updates
+[/] Rename `Keys` class to `PepelandUtilsKeyBindings`
+### 1.21.9+
+[+] Improve debug hud entries registering with logging and ability to get entries in code